test: split ACP manager turn result coverage
